split out instrumentation into its own header file
