+++ antithesis-sdk-cpp.nix
@@ -2,13 +2,14 @@
   with pkgs;
   stdenv.mkDerivation {
     pname = "antithesis-sdk-cpp";
-    version = "0.2.3";
+    version = "0.2.4";
 
     src = ./.;
 
     installPhase = ''
       mkdir -p $out/include
       cp antithesis_sdk.h $out/include/antithesis_sdk.h
+      cp antithesis_instrumentation.h $out/include/antithesis_instrumentation.h
     '';
 
   }
