+++ antithesis_sdk.h
@@ -12,7 +12,7 @@
 #include <vector>
 
 namespace antithesis {
-    inline const char* SDK_VERSION = "0.2.3";
+    inline const char* SDK_VERSION = "0.2.4";
     inline const char* PROTOCOL_VERSION = "1.0.0";
 
     struct LocalRandom {
@@ -129,19 +129,19 @@ namespace antithesis {
 
             void* fuzz_json_data = dlsym(shared_lib, "fuzz_json_data");
             if (!fuzz_json_data) {
-                error("Can access symbol fuzz_json_data");
+                error("Can not access symbol fuzz_json_data");
                 return nullptr;
             }
 
             void* fuzz_flush = dlsym(shared_lib, "fuzz_flush");
             if (!fuzz_flush) {
-                error("Can access symbol fuzz_flush");
+                error("Can not access symbol fuzz_flush");
                 return nullptr;
             }
 
             void* fuzz_get_random = dlsym(shared_lib, "fuzz_get_random");
             if (!fuzz_get_random) {
-                error("Can access symbol fuzz_get_random");
+                error("Can not access symbol fuzz_get_random");
                 return nullptr;
             }
 
@@ -576,79 +576,6 @@ namespace {
 #define SOMETIMES(cond, message, ...) ANTITHESIS_ASSERT_RAW(antithesis::SOMETIMES_ASSERTION, cond, message, __VA_ARGS__)
 #define REACHABLE(message, ...) ANTITHESIS_ASSERT_RAW(antithesis::REACHABLE_ASSERTION, true, message, __VA_ARGS__)
 #define UNREACHABLE(message, ...) ANTITHESIS_ASSERT_RAW(antithesis::UNREACHABLE_ASSERTION, false, message, __VA_ARGS__)
-
-
-// If the libvoidstar(determ) library is present, 
-// pass thru trace_pc_guard related callbacks to it
-typedef void (*trace_pc_guard_init_fn)(uint32_t *start, uint32_t *stop);
-typedef void (*trace_pc_guard_fn)(uint32_t *guard);
-
-static trace_pc_guard_init_fn trace_pc_guard_init = nullptr;
-static trace_pc_guard_fn trace_pc_guard = nullptr;
-static bool did_check_libvoidstar = false;
-static bool has_libvoidstar = false;
-
-inline void message_out(const char *msg) {
-  write(1, msg, strlen(msg));
-  return;
-}
-
-inline void load_libvoidstar() {
-    if (did_check_libvoidstar) {
-      return;
-    }
-    did_check_libvoidstar = true;
-    void* shared_lib = dlopen(antithesis::LIB_PATH, RTLD_NOW);
-    if (!shared_lib) {
-        message_out("Can not load the Antithesis native library\n");
-        return;
-    }
-
-    void* trace_pc_guard_init_sym = dlsym(shared_lib, "__sanitizer_cov_trace_pc_guard_init");
-    if (!trace_pc_guard_init_sym) {
-        message_out("Can not forward calls to libvoidstar for __sanitizer_cov_trace_pc_guard_init\n");
-        return;
-    }
-
-    void* trace_pc_guard_sym = dlsym(shared_lib, "__sanitizer_cov_trace_pc_guard");
-    if (!trace_pc_guard_sym) {
-        message_out("Can not forward calls to libvoidstar for __sanitizer_cov_trace_pc_guard\n");
-        return;
-    }
-
-    trace_pc_guard_init = reinterpret_cast<trace_pc_guard_init_fn>(trace_pc_guard_init_sym);
-    trace_pc_guard = reinterpret_cast<trace_pc_guard_fn>(trace_pc_guard_sym);
-    has_libvoidstar = true;
-}
-
-// The following symbols are indeed reserved identifiers, since we're implementing functions defined
-// in the compiler runtime. Not clear how to get Clang on board with that besides narrowly suppressing
-// the warning in this case. The sample code on the CoverageSanitizer documentation page fails this 
-// warning!
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wreserved-identifier"
-extern "C" inline void __sanitizer_cov_trace_pc_guard_init(uint32_t *start, uint32_t *stop) {
-    message_out("SDK forwarding to libvoidstar for __sanitizer_cov_trace_pc_guard_init()\n");
-    if (!did_check_libvoidstar) {
-        load_libvoidstar();
-    }
-    if (has_libvoidstar) {
-        trace_pc_guard_init(start, stop);
-    }
-    return;
-}
-
-extern "C" inline void __sanitizer_cov_trace_pc_guard( uint32_t *guard ) {
-    if (has_libvoidstar) {
-        trace_pc_guard(guard);
-    } else {
-        if (guard) {
-          *guard = 0;
-        }
-    }
-    return;
-}
-#pragma clang diagnostic pop
 
 #endif
